Convert bin/lint To ksh
Convert the `bin/lint` script to ksh so that now all development scripts for
ksh can be executed from within the ksh shell.

This also fixes an issue with the `scripts/extract_flags` script where it was
including spurious name-mangled temporary files in the `-I` arguments. Instead
file names are checked for the presence of the name-mangling `@` character, a
character which should not appear in any valid file if this is to continue
working.