Eliminate String Termination Error
Addresses Coverity #294196.

This change simply removes the `n < size` check in `path_prog`. Removing this
branch prevents coverity from taking a false path here as well as removes a
superfluous check. We then handle the scenario formerly handled by the branch
by passing `size - 1` instead of `size` to `readlink` thereby only allowing it
to fill up to one before the end of the buffer. This is the approach
recommended by CERT POS30-C for gracefully handling the behavior of `readlink`
and properly terminating its result strings.